decouple account data into separate AccountData component
